requirements.txt
alembic==1.0.8
amqp==2.5.0
aniso8601==6.0.0
billiard==3.6.0.0
blinker==1.4
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
flask-paginate==0.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.6.1
Mako==1.0.8
MarkupSafe==1.1.1
Pillow==6.0.0
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
python-memcached==1.59
pytz==2019.1
qiniu==7.2.4
redis==3.2.1
requests==2.22.0
shortuuid==0.5.0
signals==0.0.2
six==1.12.0
SQLAlchemy==1.3.2
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.1
WTForms==2.2.1
